+++ CHANGES.rst
@@ -66,6 +66,7 @@ Improvements
   :user:`duartegalvao, unconventionaldotdev`)
 - Show a warning when bulk registration approval/rejection skips registrations that are not
   pending (:issue:`7197`, :pr:`7205`, thanks :user:`duartegalvao, unconventionaldotdev`)
+- Add a JSON endpoint that returns the event's program/tracks (:pr:`7207`)
 
 Bugfixes
 ^^^^^^^^